requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
